........
  r114578 | kreese | 2015-03-16 13:07:19 -0400 (Mon, 16 Mar 2015) | 1 line
  
  Add more ignored attributes in CSR yaml
........
  r114579 | kreese | 2015-03-16 13:07:22 -0400 (Mon, 16 Mar 2015) | 1 line
  
  CSR update as of 3/16/2015. USB enum was hacked up and the NIC required lots of hacking. NIC structures now cross 64bit boundaries, hopefully this will be fixed. Only affects pass 2, so hopefully nobody cares about my hacks
........
  r114580 | kreese | 2015-03-16 13:26:29 -0400 (Mon, 16 Mar 2015) | 1 line
  
  Support linking the BDK for non-zero nodes and booting off them
........
  r114581 | kreese | 2015-03-16 13:26:32 -0400 (Mon, 16 Mar 2015) | 1 line
  
  Use the right node number for filenames when booting on non-zero nodes
........
  r114582 | kreese | 2015-03-16 13:26:36 -0400 (Mon, 16 Mar 2015) | 4 lines
  
  1) Add new QLM function bdk_qlm_rx_equalization() for performing an RX equalization evaluation.
  2) Do an RX equalization 10us after new QLM tuning parameters are applied
  3) Use the new function in the Lua QLM tuning code
  4) RX equalization is only done on 8G and 10G QLMs as it requires training support.
........
  r114583 | kreese | 2015-03-16 13:40:25 -0400 (Mon, 16 Mar 2015) | 3 lines
  
  Revert "Add CCPI 10G-KR training work-around code that detects and fixes lanes that have hung in KR training.  "
  
  This reverts commit 25b354b5f5f9ba75bf70652a2b927eea205bcc15.
........
  r114584 | kreese | 2015-03-16 13:40:27 -0400 (Mon, 16 Mar 2015) | 3 lines
  
  Revert "Work-around code for CCPI 10G links which have lanes that hang in 10G-KR training.  This file is included in the BDK's boot-stub.c, and the entry point function, run_boot_stub_ccpi(), is called in the boot-stub.c's main() function to check for stuck lanes and fix them.  "
  
  This reverts commit 05680b61a2caccf66b878c281385f1937e077ddd.
........
  r114585 | kreese | 2015-03-16 14:21:24 -0400 (Mon, 16 Mar 2015) | 1 line
  
  Enable multinode in asim networking tests
........
  r114654 | kreese | 2015-03-17 11:49:02 -0400 (Tue, 17 Mar 2015) | 2 lines
  
  Redo RX equalization, except for CCPI. CCPI doesn't recover from the
  errors generated during RX equalization.
........
  r114659 | kreese | 2015-03-17 17:34:59 -0400 (Tue, 17 Mar 2015) | 1 line
  
  Fix bug in gpio routines that didn't allow access to gpios above 32
........
  r114664 | kreese | 2015-03-17 18:16:29 -0400 (Tue, 17 Mar 2015) | 1 line
  
  Fix range check on GPIO number
........
  r114665 | kreese | 2015-03-17 18:16:30 -0400 (Tue, 17 Mar 2015) | 1 line
  
  Turn on USB power on the CRB-2S using GPIO 38
........
  r114676 | kreese | 2015-03-17 20:14:29 -0400 (Tue, 17 Mar 2015) | 1 line
  
  Add ability to repeat the SATA automated pattern test
........
  r114679 | kreese | 2015-03-17 21:06:06 -0400 (Tue, 17 Mar 2015) | 1 line
  
  Update asim sata test to account for new repeat prompt
........
  r114841 | kreese | 2015-03-19 19:39:13 -0400 (Thu, 19 Mar 2015) | 8 lines
  
  Allow BDK_LINK_ADDRESS to change without rebuilding the BDK libs.
  1) Change how the filesystem devices are initialized so BDK_LINK_ADDRESS. The device
  files for the uart are fixed up with the correct node number when __bdk_fs_init()
  is called. This must be called before any uart output.
  2) bdk-start.S uses a pointer out of a global structure to determine the link address
  instead of BDK_LINK_ADDRESS. The assembly has no way to get the link address without
  loading it from somethign in the executable due to be relocatable. We use the pointer at
  the beginning of bdk_fs_dev_uart0 to get the upper bits of the link address.
........
  r114842 | kreese | 2015-03-19 19:39:19 -0400 (Thu, 19 Mar 2015) | 1 line
  
  Calls to RX equalization on QLMs below 8G return success instead of failure
........
  r114843 | kreese | 2015-03-19 19:39:21 -0400 (Thu, 19 Mar 2015) | 2 lines
  
  Add support for CCPI software setup in function bdk_init_ccpi_links(). This works if the
  chip is strapped for software setup, but not for change speeds.
........
  r114844 | kreese | 2015-03-19 19:39:24 -0400 (Thu, 19 Mar 2015) | 2 lines
  
  Have the BDK boot stub bringup CCPI through software when it is run on a non-zero node ID
  with CCPI in software mode.
........
  r114845 | kreese | 2015-03-19 19:39:26 -0400 (Thu, 19 Mar 2015) | 1 line
  
  if CCPI is in software init mode, initialize it when bdk_init_nodes() is called. This changed the arguments to bdk_init_nodes().
........
  r114846 | kreese | 2015-03-19 19:39:32 -0400 (Thu, 19 Mar 2015) | 1 line
  
  Add make rules to generate bdk-node1.bin, used for software init of CCPI on the second node
........
  r114847 | kreese | 2015-03-19 19:39:34 -0400 (Thu, 19 Mar 2015) | 1 line
  
  Add support for software init of CCPI using the EBB8800 boot stub
........
  r114848 | kreese | 2015-03-19 19:39:36 -0400 (Thu, 19 Mar 2015) | 1 line
  
  Add support for software init of CCPI using the EBB8804 boot stub
........
  r114849 | kreese | 2015-03-19 19:39:37 -0400 (Thu, 19 Mar 2015) | 1 line
  
  Add support for software init of CCPI using the CRB-1S boot stub
........
  r114850 | kreese | 2015-03-19 19:39:39 -0400 (Thu, 19 Mar 2015) | 1 line
  
  Add support for software init of CCPI using the CRB-2S boot stub
........
  r114851 | kreese | 2015-03-19 20:20:06 -0400 (Thu, 19 Mar 2015) | 3 lines
  
  When doing software init on node 1, use GSERX_SPD instead of the gbaud to determine if 
  the CCPI QLMs are running. A chip reset doesn't stop the QLMs, but we still need to go
  through the software init sequence.
........
  r114921 | kreese | 2015-03-20 19:29:58 -0400 (Fri, 20 Mar 2015) | 1 line
  
  Make CCPI software init more robust when handling link init
........
  r115069 | kreese | 2015-03-24 14:00:35 -0400 (Tue, 24 Mar 2015) | 1 line
  
  Fix dereference of a NULL pointer if ddr_verbose is not set
........
  r115156 | kreese | 2015-03-25 19:10:35 -0400 (Wed, 25 Mar 2015) | 3 lines
  
  1) Update EBB8800 board scripting to work with one or two nodes
  2) Add flag to board info if the we're multinode
  3) Have match not consume data if it fails. This way we can match, fail, then try another match
........
  r115157 | kreese | 2015-03-25 19:10:36 -0400 (Wed, 25 Mar 2015) | 5 lines
  
  1) Add EBB8800 test file for normal boot loader, one node, full dram test
  2) Add argument parsing to board scripting so connections can be supplied on the command line. Most
  people using an MCU with a single board only need to supply the hostname of the MCU.
  3) Factor out DRAM serial tests inot a common file
  4) Factor out boot serial tests into a common file
........


git-svn-id: http://casvn/svn/octeon/bdk/branches/thunder.ddr4_rdimm@115224 c83883e8-9f28-0410-baf1-f2025811e16a

+++ libbdk-hal/bdk-gpio.h
@@ -31,7 +31,7 @@ extern int bdk_gpio_initialize(bdk_node_t node, int gpio, int is_output, int out
  *
  * @return Status of the GPIO pins
  */
-extern uint32_t bdk_gpio_read(bdk_node_t node);
+extern uint64_t bdk_gpio_read(bdk_node_t node);
 
 /**
  * GPIO Clear pin
@@ -40,7 +40,7 @@ extern uint32_t bdk_gpio_read(bdk_node_t node);
  *                   special value.
  * @param clear_mask Bit mask to indicate which bits to drive to '0'.
  */
-extern void bdk_gpio_clear(bdk_node_t node, uint32_t clear_mask);
+extern void bdk_gpio_clear(bdk_node_t node, uint64_t clear_mask);
 
 /**
  * GPIO Set pin
@@ -49,6 +49,6 @@ extern void bdk_gpio_clear(bdk_node_t node, uint32_t clear_mask);
  *                 value.
  * @param set_mask Bit mask to indicate which bits to drive to '1'.
  */
-extern void bdk_gpio_set(bdk_node_t node, uint32_t set_mask);
+extern void bdk_gpio_set(bdk_node_t node, uint64_t set_mask);
 
 /** @} */
